+++ src/field_weather.c
@@ -811,9 +811,9 @@ void FadeScreen(u8 mode, s8 delay)
 // Note: This enables blending in all windows;
 // These bits may need to be disabled later
 // (i.e if blending lighting effects using WINOBJ)
-u16 FadeScreenHardware(u8 mode, s8 delay)
+u32 FadeScreenHardware(u32 mode, s32 delay)
 {
-    u16 bldCnt = GetGpuReg(REG_OFFSET_BLDCNT) & BLDCNT_TGT2_ALL;
+    u32 bldCnt = GetGpuReg(REG_OFFSET_BLDCNT) & BLDCNT_TGT2_ALL;
     bldCnt |= BLDCNT_TGT1_ALL;
     // enable blend in all windows
     SetGpuRegBits(REG_OFFSET_WININ, WININ_WIN0_CLR | WININ_WIN1_CLR);
